I2S updates for 4.3->5.4

+++ components/codecs/inc/alac/alac_wrapper.h
@@ -12,6 +12,7 @@
 #define __ALAC_WRAPPER_H_
 
 struct alac_codec_s;
+#include <stdbool.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -22,7 +23,7 @@ struct alac_codec_s *alac_create_decoder(int magic_cookie_size, unsigned char *m
 								unsigned char *channels, unsigned int *block_size);
 void alac_delete_decoder(struct alac_codec_s *codec);
 bool alac_to_pcm(struct alac_codec_s *codec, unsigned char* input,
-				 unsigned char *output, char channels, unsigned *out_frames);
+				 unsigned char *output, char channels, long unsigned int *out_frames);
 
 #ifdef __cplusplus
 }
